+++ src/global/const.h
@@ -209,6 +209,8 @@
 
 #define RESOLUTIONS_SIZE 10
 
+#define MAX_DIFFICULTY_PRESETS 7
+
 #if _MSC_VER > 0x500
     #define strdup _strdup // fixes error about POSIX function
     #define _USE_MATH_DEFINES // makes maths.h also define M_PI